pyproject.toml
[project]
name        = "depthflow"
description = "🌊 Images to parallax effect videos"
authors     = [{name="Tremeschin"}]
version     = "0.10.0"
readme      = "readme.md"
license     = "AGPL-3.0"
requires-python = ">=3.10"
dependencies = [
    "azure-storage-blob",
    "attrs",
    "cyclopts",
    "dearlog",
    "diskcache",
    "fastapi",
    "gradio~=6.0",
    "imageio",
    "numpy",
    "pillow",
    "platformdirs",
    "pooch",
    "pydantic",
    "scipy",
    "shaderflow",
    "torch",
    "torchvision",
    "transformers",
    "python-multipart",
    "uvicorn",
    "xxhash",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
]

[project.scripts]
depthflow = "depthflow.__main__:main"
depth     = "depthflow.__main__:main"
depthflow-api = "depthflow_api.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
